backend/src/main/java/com/miapp/service/UserDetailsServiceImpl.java
package com.miapp.service;

import java.util.Collections;

import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import com.miapp.entity.Usuario;
import com.miapp.repository.UsuarioRepository;

import lombok.RequiredArgsConstructor;

@Service
@RequiredArgsConstructor
public class UserDetailsServiceImpl implements UserDetailsService {

	private final UsuarioRepository usuarioRepository;

	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
		Usuario user = usuarioRepository.findByUsername(username)
				.orElseThrow(() -> new UsernameNotFoundException("Usuario no encontrado"));

		  return org.springframework.security.core.userdetails.User.withUsername(user.getUsername())
		            .password(user.getPassword())
		            .authorities(Collections.emptyList()) // No roles or authorities
		            .build();
	}
}
